moved profile to components root
