+++ components/Profile/index.js
@@ -1,0 +1,3 @@
+import Profile from './profile'
+
+export default Profile
